requirements.txt
asgiref==3.11.1
dj-database-url==3.0.1
Django==5.2.13
django-cors-headers==4.3.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
Pillow==10.3.0
psycopg[binary]==3.2.12
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.11.0
